status and view
