+++ src/components/pages/UserDashboard/Status.tsx
@@ -1,12 +1,19 @@
 import React from "react";
-import StatusType from "./StatusType";
+//import StatusView from "./StatusView";
+import { Link } from 'react-router-dom';
 
 type StatusProp = {
     loantype: string;
     amount: string;
-    status: number;
-    view: string;
+    statustype: number;
+    statusview: number;
 };
+
+
+
+
+
+
 
 function Status(prop: StatusProp) {
     return (
@@ -14,18 +21,116 @@ function Status(prop: StatusProp) {
             <h3 className="flex justify-center items-center">
                 {prop.loantype}
             </h3>
+
             <h3 className=" flex justify-center  items-center">
                 {prop.amount}
             </h3>
+
             <div className="">
-                <StatusType status={prop.status} />
+                <StatusType status={prop.statustype} />
             </div>
 
-            <button className=" text-accentColor cursor-pointer flex justify-center items-center">
-                {prop.view}
-            </button>
+            <div> 
+               <StatusView status={prop.statusview} /> 
+            </div>
+
+
+
         </div>
     );
 }
 
 export default Status;
+
+//status view 
+function StatusView ( { status }: { status: Number }){
+    const ViewPendingApproval = (
+        <div>
+            <Link to="/loan-information">
+                <button className=" text-accentColor cursor-pointer flex justify-center items-center">
+                    view
+                </button>
+    
+            </Link>
+    
+        </div>
+    
+    );
+
+     
+    const ViewPendingContractApproval = (
+        <div>
+            <Link to="/loan-contract">
+                <button className=" text-accentColor cursor-pointer flex justify-center items-center">
+                    view
+                </button>
+            </Link>
+        </div>
+    
+    );
+
+    switch (status) {
+        case 1:{
+            return (ViewPendingApproval);
+        }
+        case 2: {
+            return (ViewPendingContractApproval);
+        }
+    
+        default:
+            return(ViewPendingApproval);
+    }
+
+
+}
+
+
+//status type
+function StatusType({ status }: { status: Number }) {
+    const PendingApproval = (
+
+        <div className="flex  items-center  space-x-2">
+            <div className="min-w-[10px] min-h-[10px]  bg-red-400 border-2 border-gray-10 rounded-full " />
+            <h3 className="max-w-[200px]">Pending Approval {"      "} </h3>
+
+        </div>
+    );
+
+    const PendingContractApproval = (
+        <div className="flex items-center space-x-2">
+            <div className="min-w-[10px] min-h-[10px] bg-red-900 border-2 border-gray-10 rounded-full " />
+            <h3 >Pending Contract Approval</h3>
+
+
+        </div>
+    );
+
+    const ViewPendingApproval = (
+        <div>
+            <Link to="/loan-information">
+                <button className=" text-accentColor cursor-pointer flex justify-center items-center">
+                    view
+                </button>
+    
+            </Link>
+    
+        </div>
+    
+    );
+
+    switch (status) {
+        case 1: {
+            return (PendingApproval);
+        }
+
+        case 2: {
+            return PendingContractApproval;
+        }
+
+        default: {
+            return PendingApproval;
+        }
+    }
+}
+
+
