+++ src/pages/LoanContract.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-
+import { Link } from 'react-router-dom';
 // layout
 import DefaultLayout from "../components/layout/DefaultLayout";
 
@@ -14,20 +14,28 @@ function LoanContract() {
                 <Header />
                 <Information />
 
+
+
+
                 <div className=" pt-5 space-x-6">
-                <button
-                    className="border text-right px-5 py-4 bg-primaryColor text-contColor rounded-lg cursor-pointer hover:shadow-sm hover:bg-accentColor duration-100 ease-in-out transition-all"
-                    type="submit"
-                >
-                    Accept
-                </button>
-                <button
-                    className="border px-5 py-4 bg-white border-black text-darkTextColor rounded-lg cursor-pointer hover:shadow-sm hover:bg-accentColor duration-100 ease-in-out transition-all"
-                    type="submit"
-                >
-                    Decline
-                </button>
-            </div>
+                    <Link to="/dashboard">
+                        <button
+                            className="border text-right px-5 py-4 bg-primaryColor text-contColor rounded-lg cursor-pointer hover:shadow-sm hover:bg-accentColor duration-100 ease-in-out transition-all"
+                            type="submit"
+                        >
+                            Accept
+                        </button>
+                    </Link>
+                    <Link to="/dashboard">
+                        <button
+                            className="border px-5 py-4 bg-white border-black text-darkTextColor rounded-lg cursor-pointer hover:shadow-sm hover:bg-accentColor duration-100 ease-in-out transition-all"
+                            type="submit"
+                        >
+                            Decline
+                        </button>
+                    </Link>
+
+                </div>
             </>
         </DefaultLayout>
     );
